scripts/requirements_gemini.txt
google-generativeai>=0.3.0



